fix: ホテルの情報の渡し方を変更
